syz-mutate: reproducible runs with no input program
If one specifies a seed, one expects a reproducible run. However, since
iteration order over a map is not stable, even with a fixed seed, the
order of syscalls will vary per run and thus when syz-mutate is called
without specifying an input program the program it generates will vary
per run.

We introduce an order by syscall ID and sort the syscall list, ensuring
that with a fixed seed, syz-mutate will generate an initial program that
is the same.

Prior to this diff, running `syz-mutate -seed 1234` would generate a
different program each time. Now it generates the same program each
time.